[devdays] Remove dependence on hbsfy
Summary:
handlebars-redux originally depended on hbsfy (for Browserify) to register helpers. This is bad, because not everyone uses grunt (i.e. this doesn't work with gulp of webpack). This resolves that issue by:
 - Moving helper initializing out of Handlebars component
 - Removing dependencies from package.json
This also fixes an issue where properties with a null value were _not_ triggering a state change as they should

Test Plan: Tested w/ a example todo app that will be put on Github shortly

Reviewers: #customer, marek.laska

Reviewed By: #customer, marek.laska

Subscribers: marek.laska, brent.foosness

Differential Revision: https://phabricator.sqr.io/D4282